# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Crosstyan",
            "email": "crosstyan@outlook.com"
        },
        "license": {
            "name": "WTFPL",
            "url": "http://www.wtfpl.net/"
        },
        "version": "0.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/humidity": {
            "get": {
                "description": "get Temperature/Humidity by page",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MQTTRecords"
                ],
                "summary": "Get Temperature/Humidity Records by Page",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "From 1 to infinity",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseMsg"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorMsg"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorMsg"
                        }
                    }
                }
            },
            "post": {
                "description": "get Temperature/Humidity by date",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MQTTRecords"
                ],
                "summary": "Get Temperature/Humidity Records by Date",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.DateRangeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseMsg"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorMsg"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorMsg"
                        }
                    }
                }
            }
        },
        "/temperature": {
            "get": {
                "description": "get Temperature/Humidity by page",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MQTTRecords"
                ],
                "summary": "Get Temperature/Humidity Records by Page",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "From 1 to infinity",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseMsg"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorMsg"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorMsg"
                        }
                    }
                }
            },
            "post": {
                "description": "get Temperature/Humidity by date",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MQTTRecords"
                ],
                "summary": "Get Temperature/Humidity Records by Date",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.DateRangeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseMsg"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorMsg"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorMsg"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.DateRangeRequest": {
            "type": "object",
            "required": [
                "start"
            ],
            "properties": {
                "end": {
                    "description": "Time RFC3339",
                    "type": "string",
                    "example": "2022-01-01T00:00:00Z"
                },
                "page": {
                    "description": "Page is from 1 to infinity",
                    "type": "integer",
                    "example": 1
                },
                "start": {
                    "description": "Time RFC3339",
                    "type": "string",
                    "example": "2020-01-01T00:00:00Z"
                }
            }
        },
        "controller.ErrorMsg": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error message",
                    "type": "string",
                    "example": "error message"
                }
            }
        },
        "controller.ResponseMsg": {
            "type": "object",
            "properties": {
                "records": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.MQTTRecord"
                    }
                }
            }
        },
        "model.MQTTRecord": {
            "type": "object",
            "properties": {
                "payload": {
                    "type": "number",
                    "example": 24.23
                },
                "timestamp": {
                    "description": "Time RFC3339",
                    "type": "string",
                    "example": "2020-01-01T00:00:00Z"
                }
            }
        }
    }
}